+++ src/lib/Banner/index.ts
@@ -1,0 +1,2 @@
+export { default as Banner } from './Banner.svelte'
+export type { BannerProps } from './banner.types.js'